added a main menu, and started working on the 2d part
for some reason the 2d player jumps and falls really slowly, and the platforms generated have collision, but are invisible, i only used the test material to see if the color of the material was the problem, but it isn't. Also, if you fall off the platform, you are able to walk back through it wall falling.